Fix bug when trying compress read-only files
All files were opened using read and write mode, even if only one was
used, which did not allow compressing files that could not be modified.
Now a flag is used to indicate how the file will be used.

There were also minor changes to the progress bar display, which now
occupies a smaller space.